Accesses the chrome history database, pulls the most recent url in browser history, scrapes it, then semantically analyses its contents.
Note, as I write this, I realize it might be more effective first to add this database to a local csv before scraping every document in the browser history.

Another important note, that there are still many things "To-do", but this is still a working product that functions to the effect mentioned above.